Delete dead members from mnesia when we know that there is nothing more in flight from them. This is a bit subtle.
To do this, we keep track of the last_pub and last_ack. However, the last_ack is different for aliases of Self than for non-aliases. For non-aliases, it is the last ack seen. However, we only want to delete the member from mnesia when the last ack _retired_ is equal to the last publish seen. Thus when we inherit a member, we set our record of the last_ack to be the last ack from our new upstream - it is guaranteed that the upstream knows the least of all members regarding the dead member and so it is the correct oracle for determining whether or not the dead member can be erased. Thus we use its last ack seen count (which can of course decrease our last ack seen count as acks may have passed through ourselves which are still in flight and are yet to reach our new upstream) and now only update the last_ack seen when we _retire_ acks. Thus when the last_ack is equal to the last_pub, it is now safe to remove the entry from mnesia.