Correct group joining mechanic.
The problem with the old version is that if we have A -> B -> C, B publishes a message which starts making its way around. Now B' joins, adding itself to mnesia leading to A -> B -> B' -> C. At this point, B dies, _before_ it sends to B' the catchup message. According to mnesia, B' is responsible for B's msgs, but B' will actually receive the catchup from A, thus having the least of all information about the messages from B - in particular, it'll crash when it receives the msg B sent before dying as it's not a msg it's seen before (which should be impossible).

The solution is to ensure that when B' joins, it has to _call_ into B, and B makes the modification to mnesia. In order to really cope with crashes, B has to send out the catchup msg _before_ it makes the write to mnesia, otherwise the same situation can occur. This is fine - if B dies and so B' has to join elsewhere, it'll ignore any catchup msgs it receives from the wrong Left. Thus B' calls B, and is prepared to deal with B dying (at which point it'll try to join again, elsewhere). One final gotcha is that B' could call into B, B makes the change to mnesia. At this point, B'' tries to call to B' to join on the RHS of B'. But B' hasn't received the catchup from B yet, and so has to tell B'' to sod off and try somewhere else.

Anyway, it seems to be bug free now.